requirements.txt
requests==2.22.0
flake8==3.7.8
Flask==1.1.1
line-bot-sdk==1.13.0
gunicorn==19.9.0
pytest==5.1.0
